+++ infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/banco/jpa/movimiento/MovimientoRepositoryAdapter.java
@@ -43,9 +43,6 @@ public class MovimientoRepositoryAdapter extends AdapterOperations<Movimiento, M
 
     @Override
     public List<Movimiento> generarReporteEntreFechas(Date inicio, Date fin) {
-        /*List<MovimientoData> listaMovimientoData =
-                repository.findByFechaMovimientoBetween(inicio.toString(),fin.toString());
-        */
         return DataMapper.converitirListaMovimientosDataAListaMovimiento(repository.findByFechaMovimientoBetween(inicio, fin));
     }
 }
