+++ infrastructure/entry-points/api-rest/src/main/java/co/com/banco/cliente/dto/ClienteDTO.java
@@ -1,0 +1,14 @@
+package co.com.banco.cliente.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class ClienteDTO {
+    private String nombres;
+    private String direccion;
+    private String telefono;
+    private String contrasena;
+    private String estado;
+}
